requirements.txt
fastapi
uvicorn
stripe
requests
python-dotenv
